+++ incident_handlerv2/resources/views/customer/sensor/index.blade.php
@@ -62,16 +62,18 @@
                 <span class="expand-icon">+</span> </a>
         </div>
     </div>
-    <div class="panel-body row">
-        <div class="col-md-12">
-            <div class="vertical-top">
-                <a onclick="showFormSensor()" class="btn btn-blue btn-icon btn-icon-standalone">
-                    <i class="fa-plus"></i>
-                    <span>Agregar sensor</span>
-                </a>
+    @if(Auth::user()->isAdmin())
+        <div class="panel-body row">
+            <div class="col-md-12">
+                <div class="vertical-top">
+                    <a onclick="showFormSensor()" class="btn btn-blue btn-icon btn-icon-standalone">
+                        <i class="fa-plus"></i>
+                        <span>Agregar sensor</span>
+                    </a>
+                </div>
             </div>
         </div>
-    </div>
+    @endif
     <div class="panel-body">
         <table class="table table-bordered table-striped" id="customer-sensors-table">
             <thead>
@@ -91,15 +93,21 @@
                     <td>{{$sensor->ipv4}}</td>
                     <td>{{$sensor->mount_point}}</td>
                     <td>
-                        {!! Form::open(array('id'=>'deleteForm-'.$sensor->id,'class' => 'form-inline', 'method' => 'DELETE', 'route' => array('sensor.destroy', $sensor->id))) !!}
-                        {!! Form::hidden('id',$sensor->id) !!}
-                        <a href="{{route('sensor.show',$sensor->id)}}"
-                           class="btn btn-info btn-sm btn-icon icon-left">
-                            Ver sensor</a>
-                        <a href="{{route('sensor.edit',$sensor->id)}}"
-                           class="btn btn-secondary btn-sm btn-icon icon-left"> Editar</a>
-                        {!! Form::button('Eliminar',['class'=>'btn btn-danger btn-sm btn-icon icon-left','onClick'=> 'onClickDelete("sensor","'.$sensor->id.'")' ]) !!}
-                        {!! Form::close() !!}
+                        @if(Auth::user()->isAdmin())
+                            {!! Form::open(array('id'=>'deleteForm-'.$sensor->id,'class' => 'form-inline', 'method' => 'DELETE', 'route' => array('sensor.destroy', $sensor->id))) !!}
+                            {!! Form::hidden('id',$sensor->id) !!}
+                            <a href="{{route('sensor.show',$sensor->id)}}"
+                               class="btn btn-info btn-sm btn-icon icon-left">
+                                Ver sensor</a>
+                            <a href="{{route('sensor.edit',$sensor->id)}}"
+                               class="btn btn-secondary btn-sm btn-icon icon-left"> Editar</a>
+                            {!! Form::button('Eliminar',['class'=>'btn btn-danger btn-sm btn-icon icon-left','onClick'=> 'onClickDelete("sensor","'.$sensor->id.'")' ]) !!}
+                            {!! Form::close() !!}
+                        @else
+                            <a href="{{route('sensor.show',$sensor->id)}}"
+                               class="btn btn-info btn-sm btn-icon icon-left">
+                                Ver sensor</a>
+                        @endif
                     </td>
                 </tr>
             @endforeach
